refactor(starknet_sequencer_node): refactor component config module name
commit-id:884760de
