+++ crates/starknet_sequencer_node/src/components.rs
@@ -12,8 +12,8 @@ use starknet_monitoring_endpoint::monitoring_endpoint::{
 };
 
 use crate::clients::SequencerNodeClients;
-use crate::config::component_execution_config::ReactiveComponentMode;
 use crate::config::node_config::SequencerNodeConfig;
+use crate::config::reactive_component_config::ReactiveComponentMode;
 use crate::version::VERSION_FULL;
 
 pub struct SequencerNodeComponents {
