+++ crates/starknet_sequencer_node/src/config/mod.rs
@@ -2,7 +2,7 @@
 mod config_test;
 
 pub mod component_config;
-pub mod component_execution_config;
 pub mod node_config;
+pub mod reactive_component_config;
 #[cfg(any(feature = "testing", test))]
 pub mod test_utils;
